requirements.txt
blinker==1.4
certifi==2021.10.8
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
python-dotenv==0.19.2
sentry-sdk==1.5.7
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
